style(*): 사용하지 않는 import문 제거 및 논리적 공백 추가

+++ src/main/java/lotto/model/EarningRate.java
@@ -7,7 +7,7 @@ public class EarningRate {
         this.money = money;
     }
 
-    public double computeEarningRate(int winnings) {
+    public double computeEarningRate(long winnings) {
         double earningRate = (winnings / (double) money) * 100.0;
         return earningRate;
     }
